+++ experimental/packages/exporter-trace-otlp-proto/src/platform/browser/index.ts
@@ -13,6 +13,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-export { OTLPExporterBrowserBase } from './OTLPExporterBrowserBase';
-export { sendWithXhr } from './util';
+export { OTLPTraceExporter } from './OTLPTraceExporter';